Add ‘cover’ theme and bootstrap to index.html
The index.html of this page is now the one page ‘cover’ theme created by the creator of bootstrap himself.

It runs on bootstrap and therefore bootstrap had to be added to the project. All the correct js and css files were added in their right places.

The site template should be fully operational in this commit and will be checked on…

leanrob.github.io